Added bounds, removed a dot, did math stuff
Added a semicircular limit on the TCP movement to mirror the
limitations of a equal seglength SCARA. The “electric fence”
conditional is somewhat finicky, as it has to push back the tcp for it
to not get stuck (see Scratch bounceOnEdge thingy). Joystick class got
an upgrade that makes leaving the throttle at a max radius possible by
moving the mouse out of the circle. This was nontrivial and required
the use of some hella trill trigga nometry, but works swimmingly. Next
things to do are dive into the inverse kinematics to visualize the
actual arm segments, and perhaps make the secondary (and currently
unused) joystick control a Z-axis with a fake shadow parallax, or
rotation or something, and allow the arm to interact with objects.
After that we’ll encapsulate the arm and start making minigames.